Crate rusoto_glacier

Amazon Glacier is a storage solution for "cold data."

Amazon Glacier is an extremely low-cost storage service that provides secure, durable, and easy-to-use storage for data backup and archival. With Amazon Glacier, customers can store their data cost effectively for months, years, or decades. Amazon Glacier also enables customers to offload the administrative burdens of operating and scaling storage to AWS, so they don't have to worry about capacity planning, hardware provisioning, data replication, hardware failure and recovery, or time-consuming hardware migrations.

Amazon Glacier is a great storage choice when low storage cost is paramount and your data is rarely retrieved. If your application requires fast or frequent access to your data, consider using Amazon S3. For more information, see Amazon Simple Storage Service (Amazon S3).

You can store any kind of data in any format. There is no maximum limit on the total amount of data you can store in Amazon Glacier.

If you are a first-time user of Amazon Glacier, we recommend that you begin by reading the following sections in the Amazon Glacier Developer Guide:

What is Amazon Glacier - This section of the Developer Guide describes the underlying data model, the operations it supports, and the AWS SDKs that you can use to interact with the service.

Getting Started with Amazon Glacier - The Getting Started section walks you through the process of creating a vault, uploading archives, creating jobs to download archives, retrieving the job output, and deleting archives.

If you're using the service, you're probably looking for GlacierClient and Glacier.

Structs

AbortMultipartUploadInput
Provides options to abort a multipart upload identified by the upload ID.

For information about the underlying REST API, see Abort Multipart Upload. For conceptual information, see Working with Archives in Amazon Glacier.

AbortVaultLockInput
The input values for AbortVaultLock.

AddTagsToVaultInput
The input values for AddTagsToVault.

ArchiveCreationOutput
Contains the Amazon Glacier response to your request.

For information about the underlying REST API, see Upload Archive. For conceptual information, see Working with Archives in Amazon Glacier.

CSVInput
Contains information about the comma-separated value (CSV) file to select from.

CSVOutput
Contains information about the comma-separated value (CSV) file that the job results are stored in.

CompleteMultipartUploadInput
Provides options to complete a multipart upload operation. This informs Amazon Glacier that all the archive parts have been uploaded and Amazon Glacier can now assemble the archive from the uploaded parts. After assembling and saving the archive to the vault, Amazon Glacier returns the URI path of the newly created archive resource.

CompleteVaultLockInput
The input values for CompleteVaultLock.

CreateVaultInput
Provides options to create a vault.

CreateVaultOutput
Contains the Amazon Glacier response to your request.

DataRetrievalPolicy
Data retrieval policy.

DataRetrievalRule
Data retrieval policy rule.

DeleteArchiveInput
Provides options for deleting an archive from an Amazon Glacier vault.

DeleteVaultAccessPolicyInput
DeleteVaultAccessPolicy input.

DeleteVaultInput
Provides options for deleting a vault from Amazon Glacier.

DeleteVaultNotificationsInput
Provides options for deleting a vault notification configuration from an Amazon Glacier vault.

DescribeJobInput
Provides options for retrieving a job description.

DescribeVaultInput
Provides options for retrieving metadata for a specific vault in Amazon Glacier.

DescribeVaultOutput
Contains the Amazon Glacier response to your request.

Encryption
Contains information about the encryption used to store the job results in Amazon S3.

GetDataRetrievalPolicyInput
Input for GetDataRetrievalPolicy.

GetDataRetrievalPolicyOutput
Contains the Amazon Glacier response to the GetDataRetrievalPolicy request.

GetJobOutputInput
Provides options for downloading output of an Amazon Glacier job.

GetJobOutputOutput
Contains the Amazon Glacier response to your request.

GetVaultAccessPolicyInput
Input for GetVaultAccessPolicy.

GetVaultAccessPolicyOutput
Output for GetVaultAccessPolicy.

GetVaultLockInput
The input values for GetVaultLock.

GetVaultLockOutput
Contains the Amazon Glacier response to your request.

GetVaultNotificationsInput
Provides options for retrieving the notification configuration set on an Amazon Glacier vault.

GetVaultNotificationsOutput
Contains the Amazon Glacier response to your request.

GlacierClient
A client for the Amazon Glacier API.

GlacierJobDescription
Contains the description of an Amazon Glacier job.

Grant
Contains information about a grant.

Grantee
Contains information about the grantee.

InitiateJobInput
Provides options for initiating an Amazon Glacier job.

InitiateJobOutput
Contains the Amazon Glacier response to your request.

InitiateMultipartUploadInput
Provides options for initiating a multipart upload to an Amazon Glacier vault.

InitiateMultipartUploadOutput
The Amazon Glacier response to your request.

InitiateVaultLockInput
The input values for InitiateVaultLock.

InitiateVaultLockOutput
Contains the Amazon Glacier response to your request.

InputSerialization
Describes how the archive is serialized.

InventoryRetrievalJobDescription
Describes the options for a range inventory retrieval job.

InventoryRetrievalJobInput
Provides options for specifying a range inventory retrieval job.

JobParameters
Provides options for defining a job.

ListJobsInput
Provides options for retrieving a job list for an Amazon Glacier vault.

ListJobsOutput
Contains the Amazon Glacier response to your request.

ListMultipartUploadsInput
Provides options for retrieving list of in-progress multipart uploads for an Amazon Glacier vault.

ListMultipartUploadsOutput
Contains the Amazon Glacier response to your request.

ListPartsInput
Provides options for retrieving a list of parts of an archive that have been uploaded in a specific multipart upload.

ListPartsOutput
Contains the Amazon Glacier response to your request.

ListProvisionedCapacityInput

ListProvisionedCapacityOutput

ListTagsForVaultInput
The input value for ListTagsForVaultInput.

ListTagsForVaultOutput
Contains the Amazon Glacier response to your request.

ListVaultsInput
Provides options to retrieve the vault list owned by the calling user's account. The list provides metadata information for each vault.

ListVaultsOutput
Contains the Amazon Glacier response to your request.

OutputLocation
Contains information about the location where the select job results are stored.

OutputSerialization
Describes how the select output is serialized.

PartListElement
A list of the part sizes of the multipart upload.

ProvisionedCapacityDescription
The definition for a provisioned capacity unit.

PurchaseProvisionedCapacityInput

PurchaseProvisionedCapacityOutput

RemoveTagsFromVaultInput
The input value for RemoveTagsFromVaultInput.

S3Location
Contains information about the location in Amazon S3 where the select job results are stored.

SelectParameters
Contains information about the parameters used for a select.

SetDataRetrievalPolicyInput
SetDataRetrievalPolicy input.

SetVaultAccessPolicyInput
SetVaultAccessPolicy input.

SetVaultNotificationsInput
Provides options to configure notifications that will be sent when specific events happen to a vault.

UploadArchiveInput
Provides options to add an archive to a vault.

UploadListElement
A list of in-progress multipart uploads for a vault.

UploadMultipartPartInput
Provides options to upload a part of an archive in a multipart upload operation.

UploadMultipartPartOutput
Contains the Amazon Glacier response to your request.

VaultAccessPolicy
Contains the vault access policy.

VaultLockPolicy
Contains the vault lock policy.

VaultNotificationConfig
Represents a vault's notification configuration.

Enums

AbortMultipartUploadError
Errors returned by AbortMultipartUpload

AbortVaultLockError
Errors returned by AbortVaultLock

AddTagsToVaultError
Errors returned by AddTagsToVault

CompleteMultipartUploadError
Errors returned by CompleteMultipartUpload

CompleteVaultLockError
Errors returned by CompleteVaultLock

CreateVaultError
Errors returned by CreateVault

DeleteArchiveError
Errors returned by DeleteArchive

DeleteVaultAccessPolicyError
Errors returned by DeleteVaultAccessPolicy

DeleteVaultError
Errors returned by DeleteVault

DeleteVaultNotificationsError
Errors returned by DeleteVaultNotifications

DescribeJobError
Errors returned by DescribeJob

DescribeVaultError
Errors returned by DescribeVault

GetDataRetrievalPolicyError
Errors returned by GetDataRetrievalPolicy

GetJobOutputError
Errors returned by GetJobOutput

GetVaultAccessPolicyError
Errors returned by GetVaultAccessPolicy

GetVaultLockError
Errors returned by GetVaultLock

GetVaultNotificationsError
Errors returned by GetVaultNotifications

InitiateJobError
Errors returned by InitiateJob

InitiateMultipartUploadError
Errors returned by InitiateMultipartUpload

InitiateVaultLockError
Errors returned by InitiateVaultLock

ListJobsError
Errors returned by ListJobs

ListMultipartUploadsError
Errors returned by ListMultipartUploads

ListPartsError
Errors returned by ListParts

ListProvisionedCapacityError
Errors returned by ListProvisionedCapacity

ListTagsForVaultError
Errors returned by ListTagsForVault

ListVaultsError
Errors returned by ListVaults

PurchaseProvisionedCapacityError
Errors returned by PurchaseProvisionedCapacity

RemoveTagsFromVaultError
Errors returned by RemoveTagsFromVault

SetDataRetrievalPolicyError
Errors returned by SetDataRetrievalPolicy

SetVaultAccessPolicyError
Errors returned by SetVaultAccessPolicy

SetVaultNotificationsError
Errors returned by SetVaultNotifications

UploadArchiveError
Errors returned by UploadArchive

UploadMultipartPartError
Errors returned by UploadMultipartPart

Traits

Glacier
Trait representing the capabilities of the Amazon Glacier API. Amazon Glacier clients implement this trait.