Struct rusoto_glacier::DescribeVaultOutput

pub struct DescribeVaultOutput {
    pub creation_date: Option<String>,
    pub last_inventory_date: Option<String>,
    pub number_of_archives: Option<i64>,
    pub size_in_bytes: Option<i64>,
    pub vault_arn: Option<String>,
    pub vault_name: Option<String>,
}

Contains the Amazon Glacier response to your request.

Fields
creation_date: Option<String>

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

last_inventory_date: Option<String>

The Universal Coordinated Time (UTC) date when Amazon Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

number_of_archives: Option<i64>

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

size_in_bytes: Option<i64>

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

vault_arn: Option<String>

The Amazon Resource Name (ARN) of the vault.

vault_name: Option<String>

The name of the vault.