Struct rusoto_glacier::GlacierClient

pub struct GlacierClient { /* fields omitted */ }

A client for the Amazon Glacier API.

Methods

impl GlacierClient

pub fn new(region: Region) -> GlacierClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> GlacierClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,