Struct rusoto_glacier::GlacierJobDescription

pub struct GlacierJobDescription {
    pub action: Option<String>,
    pub archive_id: Option<String>,
    pub archive_sha256_tree_hash: Option<String>,
    pub archive_size_in_bytes: Option<i64>,
    pub completed: Option<bool>,
    pub completion_date: Option<String>,
    pub creation_date: Option<String>,
    pub inventory_retrieval_parameters: Option<InventoryRetrievalJobDescription>,
    pub inventory_size_in_bytes: Option<i64>,
    pub job_description: Option<String>,
    pub job_id: Option<String>,
    pub retrieval_byte_range: Option<String>,
    pub sha256_tree_hash: Option<String>,
    pub sns_topic: Option<String>,
    pub status_code: Option<String>,
    pub status_message: Option<String>,
    pub tier: Option<String>,
    pub vault_arn: Option<String>,
}

Describes an Amazon Glacier job.

Fields
action: Option<String>

The job type. It is either ArchiveRetrieval or InventoryRetrieval.

archive_id: Option<String>

For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

archive_sha256_tree_hash: Option<String>

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

archive_size_in_bytes: Option<i64>

For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

completed: Option<bool>

The job status. When a job is completed, you get the job's output.

completion_date: Option<String>

The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

creation_date: Option<String>

The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

inventory_retrieval_parameters: Option<InventoryRetrievalJobDescription>

Parameters used for range inventory retrieval.

inventory_size_in_bytes: Option<i64>

For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

job_description: Option<String>

The job description you provided when you initiated the job.

job_id: Option<String>

An opaque string that identifies an Amazon Glacier job.

retrieval_byte_range: Option<String>

The retrieved byte range for archive retrieval jobs in the form "StartByteValue-EndByteValue" If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

sha256_tree_hash: Option<String>

For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null.

The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value.

For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null in the following situations:

Archive retrieval jobs that specify a range that is not tree-hash aligned.

Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress.

Inventory jobs.

sns_topic: Option<String>

An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

status_code: Option<String>

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

status_message: Option<String>

A friendly message that describes the job status.

tier: Option<String>

The retrieval option to use for the archive retrieval. Valid values are Expedited, Standard, or Bulk. Standard is the default.

vault_arn: Option<String>

The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.